:art chore/refactor
cria os services AvailalbleService, CancelAppointmentService, CreateAppointment Service

Move parte da lógica dos controllers AppointmentController, AvalilableController, SessionController, UserController para os services

cria helpers/ErrorHelper
