+++ src/app.js
@@ -39,12 +39,16 @@ class App {
 
   exceptionHandler() {
     this.server.use(async (err, req, res, next) => {
+      const errors = await new Youch(err, req).toJSON();
+      const { status } = errors.error;
+
       if (process.env.NODE_ENV === 'development') {
-        const errors = await new Youch(err, req).toJSON();
-        return res.status(500).json(errors);
+        return res.status(status || 500).json(errors);
       }
 
-      return res.status(500).json({ error: 'Internal server error' });
+      return res.status(status || 500).json({
+        error: status === 500 ? 'Internal server error' : errors.error.message,
+      });
     });
   }
 }
