+++ src/app/helpers/ErrorHelper.js
@@ -1,0 +1,8 @@
+class ErrorHelper extends Error {
+  constructor(status, ...args) {
+    super(args);
+    this.status = status;
+  }
+}
+
+export default ErrorHelper;
